refactor: Modal 클래스 분리 : Modal, MovieInfo, VoteHandler
